Move shared code in ex18{,p} to a "header" file
